feat(GTK): Support for some GTK TextBox visual properties, clear button
- Some visual properties of TextBox now properly flow to native
- Clear button works
