+++ src/Uno.UI.Runtime.Skia.Gtk/GtkApplicationViewExtension.cs
