+++ src/Uno.UI.Runtime.Skia.Gtk/GtkCoreWindowExtension.cs
