+++ src/Uno.UI.Runtime.Skia.Gtk/Helpers/Theming/GtkSystemThemeHelperExtension.cs
